requirements.txt
opencv-python-headless
pillow
diffusers
fiftyone==0.23.0rc1
transformers
accelerate
scipy
xformers
torch
torchvision
torchaudio
